+++ editors/atlas-grounding-editor/components/GroundingForm.tsx
@@ -7,7 +7,6 @@ import {
   getTagText,
 } from "../../shared/utils/utils.js";
 import { type PHIDOption } from "@powerhousedao/document-engineering/ui";
-import type { EditorMode } from "../../shared/types.js";
 import { getOriginalNotionDocument } from "../../../document-models/utils.js";
 import { type ParsedNotionDocumentType } from "../../../scripts/apply-changes/atlas-base/NotionTypes.js";
 import { FormModeProvider } from "../../shared/providers/FormModeProvider.js";
@@ -23,14 +22,14 @@ import { DocTypeForm } from "../../shared/components/forms/DocTypeForm.js";
 import { MasterStatusForm } from "../../shared/components/forms/MasterStatusForm.js";
 import { SinglePhIdForm } from "../../shared/components/forms/SinglePhIdForm.js";
 import { GlobalTagsForm } from "../../shared/components/forms/GlobalTagsForm.js";
-import { useCallback, useEffect, useState } from "react";
+import { useState } from "react";
 import {
   getFlexLayoutClassName,
   getWidthClassName,
 } from "../../shared/utils/styles.js";
-import { MarkdownEditor } from "../../shared/components/markdown-editor.js";
 import { MultiPhIdForm } from "../../shared/components/forms/MultiPhIdForm.js";
 import { useParentOptions } from "../../shared/hooks/useParentOptions.js";
+import { MarkdownContentForm } from "../../shared/components/forms/MarkdownContentForm.js";
 
 interface GroundingFormProps extends Pick<IProps, "document" | "dispatch"> {
   mode: ViewMode;
@@ -63,26 +62,6 @@ export function GroundingForm({
   const documentState = {
     ...originalDocumentState,
     parent: parentId,
-  };
-
-  const [contentValue, setContentValue] = useState(documentState.content ?? "");
-
-  // Update contentValue when documentState changes
-  useEffect(() => {
-    setContentValue(documentState.content ?? "");
-  }, [documentState.content]);
-
-  // Custom handler for content changes
-  const handleContentChange = (value: string) => {
-    setContentValue(value);
-  };
-
-  // Custom handler for content blur
-  const handleContentBlur = () => {
-    // Only save if the content has actually changed
-    if (contentValue !== documentState.content) {
-      dispatch(actions.setContent({ content: contentValue }));
-    }
   };
 
   // baseline node state
@@ -152,12 +131,11 @@ export function GroundingForm({
             </div>
           </div>
           <div className={cn("flex-1 min-h-[350px]")}>
-            <MarkdownEditor
-              value={contentValue}
-              onChange={handleContentChange}
-              onBlur={handleContentBlur}
-              height={350}
-              label="Content"
+            <MarkdownContentForm
+              value={documentState.content ?? ""}
+              onSave={(value) => {
+                dispatch(actions.setContent({ content: value }));
+              }}
             />
           </div>
